Refactor `teos::api::http` tests to reuse code for lightning tests
